# server/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "嘉兴项目接口",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/flow/node": {
      "get": {
        "tags": ["审批"],
        "summary": "查询当前、下一节点信息",
        "description": "查询当前、下一节点信息",
        "parameters": [
          {
            "name": "idAuditing",
            "in": "query",
            "description": "审核流程ID",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queryType",
            "in": "query",
            "description": "1:当前节点 2:下一节点",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "msg": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}